Split declaration validation resolver replay tests
